perception/visual/minicpm_requirements.txt
accelerate==0.30.1
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==1.3.1
alembic==1.11.1
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.1
altair==5.3.0
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.6.2
blis==0.7.11
catalogue==2.0.10
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.6
cloudpathlib==0.16.0
cloudpickle==2.2.1
cmake==3.25.0
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
crcmod==1.7
croniter==1.3.15
cryptography==41.0.3
cycler==0.12.1
cymem==2.0.8
databricks-cli==0.17.7
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.3.1
diskcache==5.6.3
docker==6.1.3
docker-pycreds==0.4.0
editdistance==0.6.2
einops==0.7.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
fairscale==0.4.0
fastapi==0.110.3
ffmpy==0.3.2
filelock==3.14.0
Flask==2.3.2
fonttools==4.52.4
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
gradio==4.26.0
gradio_client==0.15.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.4
imageio==2.31.1
importlib_resources==6.4.0
inquirer==3.1.3
interegular==0.3.3
ipdb==0.13.13
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.1
jsonlines==4.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lark==1.1.9
lazy_loader==0.3
lightning==2.0.1
lightning-cloud==0.5.37
lightning-utilities==0.9.0
lit==15.0.7
llvmlite==0.42.0
lm-format-enforcer==0.9.8
lxml==5.2.2
Mako==1.2.4
marisa-trie==1.1.1
Markdown==3.4.4
markdown-it-py==3.0.0
markdown2==2.4.10
MarkupSafe==2.1.2
mat4py==0.5.0
matplotlib==3.7.4
matplotlib-inline==0.1.6
mdurl==0.1.2
model-index==0.1.11
more-itertools==10.1.0
mpmath==1.2.1
msgpack==1.0.8
multidict==6.0.4
murmurhash==1.0.10
nest-asyncio==1.6.0
networkx==3.0
ninja==1.11.1.1
nltk==3.8.1
numba==0.59.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python-headless==4.5.5.64
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.2
openxlab==0.0.15
ordered-set==4.1.0
orjson==3.10.3
oss2==2.17.0
outlines==0.0.34
packaging==23.2
pandas==2.0.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
portalocker==2.8.2
preshed==3.0.9
prettytable==3.8.0
prometheus_client==0.20.0
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pycocotools==2.0.6
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.7.2
pydantic_core==2.18.3
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.8.0
pynvml==11.5.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-editor==1.0.4
python-multipart==0.0.9
pytorch-lightning==2.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
querystring-parser==1.2.4
ray==2.23.0
readchar==4.0.5
redis==4.6.0
referencing==0.35.1
regex==2023.6.3
requests==2.28.2
rich==13.4.2
rpds-py==0.18.1
ruff==0.4.6
sacrebleu==2.3.2
safetensors==0.4.3
scikit-image==0.21.0
scipy==1.11.1
seaborn==0.13.0
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.29.2
setproctitle==1.3.2
shapely==2.0.1
shellingham==1.5.4
shortuuid==1.0.11
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.4.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.19
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.2
starlette==0.37.2
starsessions==1.3.0
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboardX==2.6.2
terminaltables==3.1.10
thinc==8.2.3
threadpoolctl==3.2.0
tifffile==2023.7.18
tiktoken==0.6.0
timm==0.9.10
tokenizers==0.19.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.2
torchmetrics==1.0.1
torchvision==0.16.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.40.0
triton==2.1.0
typer==0.9.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.13
uvicorn==0.24.0.post1
uvloop==0.19.0
# Editable install with no version control (vllm==0.4.1)
-e /data1/zhangxin/llm/code/MiniCPM-V/vllm
vllm-nccl-cu12==2.18.1.0.4.0
wasabi==1.1.2
watchfiles==0.22.0
wcwidth==0.2.6
weasel==0.3.4
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
xformers==0.0.25
yarl==1.9.2
